requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib 
transformers 
torch 
torchvision 
faiss-cpu
Pillow 
numpy 
requests